Add an S3 print method for prettier console output
This commit illustrates how the return list used for api calls
can be assigned a simple S3 class `github`, such that S3 methods
(any function ending in `.github`) can be added for this class.

I illustrate this with a print method which formats the output
of an API return call in a slightly nicer way. Note that my
output function introduces a dependency on the `yaml` package,
available on CRAN.